reformatting modules
